requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer
cryptography==38.0.4
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.4
django-allauth==0.51.0
django-bootstrap5==22.2
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-pwa-webpush==1.1.3
django-sass-compiler==1.1.0
django-simple-pwa==2.2.1
fontawesomefree==6.2.1
gunicorn==20.1.0
http-ece==1.1.0
idna==3.4
libsass==0.22.0
oauthlib==3.2.2
Pillow==9.3.0
psycopg2==2.9.5
py-vapid==1.9.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pywebpush==1.14.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
urllib3==1.26.13
whitenoise==6.2.0
